assistant-carriere/requirements.txt
streamlit
requests
python-dotenv
pypdf
pandas
pdfplumber